Controller refactoring - roles refactoring for better reuse in API - reviewd API base class - let objectform based controllers be used by APIs or partial update - first API controllers for building/rack/device - removed form prefix - API added message for 401 status

+++ lib/App/Manoc/CatalystRole/Permission.pm
@@ -193,6 +193,8 @@ check_permission or detach to access denied page.
 sub require_permission {
     my $c = shift;
 
+    return if $c->stash->{skip_permission_check};
+
     $c->check_permission(@_) or
         $c->detach('/access_denied');
 }
